[hotfix][table] Move JoinedRowData to utils subpackage
